[www] Make clicking card categories set filters (#6774)
* clicking on categories on cards now makes a query for them

* Refactor showcase item category links to component

… so we can use it for the „Featured Sites“ card categories, too.

* Link „Featured Sites“ categories

* Move „Featured Sites“ to component

* Remove whitespace

* Make „Featured Sites“ category link scroll to <FilteredShowcase>

Move „Featured Sites“ styles around, too. Missed that in 4582fc4.

* Bring back Futura PT for „Featured Sites“ „built by“
